implemented UI change of book chapter popup menu for desktop and minor fixes
