--- src/components/Dropdown.tsx
@@ -1,28 +1,0 @@
-import Select from "react-select";
-import { DropdownProps } from "../types/Navigation";
-
-const Dropdown = ({
-  options,
-  value,
-  onChange,
-  placeholder,
-  formatOptionLabel,
-  zIndex = "z-10",
-}: DropdownProps) => {
-  return (
-    <div className={`relative ${zIndex}`}>
-      <Select
-        options={options}
-        value={value}
-        formatOptionLabel={formatOptionLabel}
-        onChange={onChange}
-        placeholder={placeholder}
-        isOptionDisabled={(option) =>
-          ("isDisabled" in option && option.isDisabled) ?? false
-        }
-      />
-    </div>
-  );
-};
-
-export default Dropdown;
